+++ affected-paths/core/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Affected-Paths Core
 
 ## Unreleased
+- Adds `autoInjectPlugin` flag to `CoreOptions`, which auto-injects the "com.squareup.tooling" plugin to the build
 
 ## v0.1.0
 - Initial public release